users/serializers/user_profile_serializer.py
from rest_framework import serializers

from users.models import UserProfile


class UserProfileSerializer(serializers.ModelSerializer):
    username = serializers.CharField(source="user.username", read_only=True)

    class Meta:
        model = UserProfile
        read_only_fields = ("username", "uuid")
        fields = read_only_fields

