pyproject.toml
[tool.poetry]
name = "bitpin-interview-task"
version = "0.1.0"
description = ""
authors = ["mammrezesk <mohammadreza.eskandari@divar.ir>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.7"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
uuid = "^1.30"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

